Add(minjae/image): detail image slide 리팩토링 #83
